pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[job-queue]>=22.6",
    "flask>=3.0.0",
    "requests>=2.32.5",
    "beautifulsoup4>=4.12.0",
]
